+++ src/gap_instance.h
@@ -20,8 +20,8 @@ class GapInstance
         int n; // Número de vendedores.
         int maxCost; // c_max definido en la consigna.
         std::vector<std::vector<int>> demands; // matriz m x n: linea = deposito, columna = vendedor. pos ij = demanda del vendedor j al deposito i.
-        std::vector<std::vector<int>> costs; // matriz m x n: linea = deposito, columna = vendedor. pos ij = costo de asignar el vendedor j al deposito i.
-        std::vector<int> capacities; // vector de tamaño m. pos i = capacidad del deposito i.
+        std::vector<std::vector<double>> costs; // matriz m x n: linea = deposito, columna = vendedor. pos ij = costo de asignar el vendedor j al deposito i.
+        std::vector<double> capacities; // vector de tamaño m. pos i = capacidad del deposito i.
 
 };
